requirements.txt
bcrypt==4.0.1
beanie==1.29.0
dynaconf==3.1.8
fastapi==0.115.8
httpcore==0.16.3
httpx==0.23.1
lark==1.2.2
motor==3.1.1
orjson==3.9.14
openpyxl==3.1.2
passlib==1.7.4
pycparser==2.21
pydantic==2.7.4
pydantic_settings==2.7.1
PyJWT==2.0.0
pymongo==4.8.0
python-dotenv==0.21.0
python-multipart==0.0.5
starlette==0.40.0
typing_extensions==4.12.0
ujson==5.6.0
uvicorn==0.20.0
watchfiles==0.22
websockets==10.4
slowapi==0.1.8

# For ML
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.5
xgboost==2.1.4